test(Assets): consolidate getAgentLinkForConversation tests into utils.spec.ts and update duplicate role action test (Issue #4332, #4335, #4337)
